!IB (CE-19515) (3DEngine) Fixed sys-spec switch problems with shadow caches !IB (CE-20054) (3DEngine) Fixed game-mode switch problems with shadow caches !IB (CE-21035) (Renderer, XBox) Fixed 11.1 ClearUAV !IB (CE-21125) (3DEngine) Fixed linked type assert !IB (CE-21245) (CE-19571) (Renderer, DX12) Fixed spec-change problem
Copied from Perforce
 Change: 2008628

+++ Code/CryEngine/RenderDll/Common/Textures/PlanningTextureStreamer.cpp
@@ -279,7 +279,6 @@ void CPlanningTextureStreamer::ApplySchedule(EApplyScheduleFlags asf)
 		}
 
 		CTexture::s_nStatsAllocFails = m_nStreamAllocFails;
-
 	}
 	else
 	{
@@ -622,7 +621,7 @@ void CPlanningTextureStreamer::SyncWithJob_Locked() threadsafe
 }
 
 #if defined(TEXSTRM_TEXTURECENTRIC_MEMORY)
-bool CPlanningTextureStreamer::TrimTexture(int nBias, TStreamerTextureVec& trimmable, STexPool* pPrioritise)
+bool CPlanningTextureStreamer::TrimTexture(int16 nBias, TStreamerTextureVec& trimmable, STexPool* pPrioritise)
 {
 	FUNCTION_PROFILER_RENDERER();
